refactor: QueryService 를 service.query 패키지로 이동
