+++ src/main/java/com/codeit/team4/deokhugam/dashboard/service/DashboardBatchService.java
@@ -12,6 +12,9 @@ import com.codeit.team4.deokhugam.dashboard.model.PowerUserSearchModel;
 import com.codeit.team4.deokhugam.dashboard.repository.PopularBookRepository;
 import com.codeit.team4.deokhugam.dashboard.repository.PopularReviewRepository;
 import com.codeit.team4.deokhugam.dashboard.repository.PowerUserRepository;
+import com.codeit.team4.deokhugam.dashboard.service.aggregator.PopularBookAggregator;
+import com.codeit.team4.deokhugam.dashboard.service.aggregator.PopularReviewAggregator;
+import com.codeit.team4.deokhugam.dashboard.service.aggregator.PowerUserAggregator;
 import com.codeit.team4.deokhugam.notification.event.ReviewRankedEvent;
 import com.codeit.team4.deokhugam.review.entity.Review;
 import com.codeit.team4.deokhugam.review.service.ReviewService;
