+++ src/main/java/com/codeit/team4/deokhugam/dashboard/service/DashboardFacade.java
@@ -9,6 +9,9 @@ import com.codeit.team4.deokhugam.dashboard.model.PopularBookViewModel;
 import com.codeit.team4.deokhugam.dashboard.model.PopularReviewViewModel;
 import com.codeit.team4.deokhugam.dashboard.model.PowerUserViewModel;
 import com.codeit.team4.deokhugam.dashboard.model.RankedViewModel;
+import com.codeit.team4.deokhugam.dashboard.service.reader.PopularBookReader;
+import com.codeit.team4.deokhugam.dashboard.service.reader.PopularReviewReader;
+import com.codeit.team4.deokhugam.dashboard.service.reader.PowerUserReader;
 import com.codeit.team4.deokhugam.global.response.PageResponse;
 import java.time.Instant;
 import java.time.LocalDate;
