+++ src/main/java/com/codeit/team4/deokhugam/dashboard/service/aggregator/PowerUserAggregator.java
@@ -1,12 +1,11 @@
-package com.codeit.team4.deokhugam.dashboard.service;
+package com.codeit.team4.deokhugam.dashboard.service.aggregator;
 
-import static com.codeit.team4.deokhugam.jooq.tables.Books.BOOKS;
 import static com.codeit.team4.deokhugam.jooq.tables.Reviews.REVIEWS;
 import static com.codeit.team4.deokhugam.jooq.tables.Users.USERS;
 
-import com.codeit.team4.deokhugam.dashboard.builder.PopularReviewAggregationQueryBuilder;
+import com.codeit.team4.deokhugam.dashboard.builder.PowerUserAggregationQueryBuilder;
 import com.codeit.team4.deokhugam.dashboard.entity.PeriodType;
-import com.codeit.team4.deokhugam.dashboard.model.PopularReviewSearchModel;
+import com.codeit.team4.deokhugam.dashboard.model.PowerUserSearchModel;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.UUID;
@@ -19,29 +18,29 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
-public class PopularReviewAggregator {
+public class PowerUserAggregator {
 
-    private static final int POPULAR_REVIEW_LIMIT = 20;
+    private static final int POWER_USER_LIMIT = 10;
 
     private final DSLContext dsl;
-    private final PopularReviewAggregationQueryBuilder aggregationQueryBuilder;
+    private final PowerUserAggregationQueryBuilder aggregationQueryBuilder;
 
-    public List<PopularReviewSearchModel> findTopReviews(PeriodType period, LocalDate snapshotDate) {
+    public List<PowerUserSearchModel> findTopPowerUsers(PeriodType period, LocalDate snapshotDate) {
         Table<?> periodLikes = aggregationQueryBuilder.buildPeriodLikeCountTable(period, snapshotDate);
         Table<?> periodComments = aggregationQueryBuilder.buildPeriodCommentCountTable(period, snapshotDate);
 
-        return dsl.select(PopularReviewSearchModel.toSelectedFields())
+        return dsl.select(PowerUserSearchModel.toSelectedFields())
                 .from(REVIEWS)
-                .join(BOOKS).on(REVIEWS.BOOK_ID.eq(BOOKS.ID))
                 .join(USERS).on(REVIEWS.USER_ID.eq(USERS.ID))
                 .leftJoin(periodLikes).on(periodLikes.field("review_id", UUID.class).eq(REVIEWS.ID))
                 .leftJoin(periodComments).on(periodComments.field("review_id", UUID.class).eq(REVIEWS.ID))
                 .where(aggregationQueryBuilder.buildCondition(period, snapshotDate))
+                .groupBy(PowerUserSearchModel.toGroupByFields())
                 .orderBy(aggregationQueryBuilder.buildOrderBy(
-                        PopularReviewSearchModel.PERIOD_LIKE_COUNT,
-                        PopularReviewSearchModel.PERIOD_COMMENT_COUNT
+                        PowerUserSearchModel.PERIOD_LIKE_COUNT,
+                        PowerUserSearchModel.PERIOD_COMMENT_COUNT
                 ))
-                .limit(POPULAR_REVIEW_LIMIT)
-                .fetch(PopularReviewSearchModel::fromRecord);
+                .limit(POWER_USER_LIMIT)
+                .fetch(PowerUserSearchModel::fromRecord);
     }
 }
