+++ src/main/java/com/codeit/team4/deokhugam/dashboard/service/reader/PopularBookReader.java
@@ -1,4 +1,4 @@
-package com.codeit.team4.deokhugam.dashboard.service;
+package com.codeit.team4.deokhugam.dashboard.service.reader;
 
 import static com.codeit.team4.deokhugam.jooq.tables.PopularBooks.POPULAR_BOOKS;
