+++ src/test/java/com/codeit/team4/deokhugam/review/controller/ReviewControllerTest.java
@@ -22,7 +22,7 @@ import com.codeit.team4.deokhugam.review.dto.ReviewLikeResponse;
 import com.codeit.team4.deokhugam.review.dto.ReviewResponse;
 import com.codeit.team4.deokhugam.review.dto.ReviewSearchRequestParam;
 import com.codeit.team4.deokhugam.review.dto.ReviewUpdateRequest;
-import com.codeit.team4.deokhugam.review.service.ReviewQueryService;
+import com.codeit.team4.deokhugam.review.service.query.ReviewQueryService;
 import com.codeit.team4.deokhugam.review.service.ReviewService;
 import com.codeit.team4.deokhugam.user.service.UserService;
 import com.fasterxml.jackson.databind.ObjectMapper;
